Conflicts:
	src/frontend/org/voltdb/compiler/VoltCompiler.java
	src/frontend/org/voltdb/utils/CatalogUtil.java

+++ src/ee/structures/CompactingMap.h
@@ -1,5 +1,5 @@
 /* This file is part of VoltDB.
- * Copyright (C) 2008-2013 VoltDB Inc.
+ * Copyright (C) 2008-2014 VoltDB Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -194,7 +194,7 @@ template<typename Key, typename Data, typename Compare, bool hasRank>
 CompactingMap<Key, Data, Compare, hasRank>::CompactingMap(bool unique, Compare comper)
     : m_count(0),
       m_root(&NIL),
-      m_allocator(sizeof(TreeNode) - (hasRank ? 0 : sizeof(NodeCount)), 10000),
+      m_allocator(static_cast<int>(sizeof(TreeNode) - (hasRank ? 0 : sizeof(NodeCount))), static_cast<int>(10000)),
       m_unique(unique),
       m_comper(comper)
   {
